Features addition - Problem solved
- Link portfolio sudah saya ganti menjadi link GIT saya.
- Geser-geser sedikit layout text pada scene Main Menu dan Credit.
- Perbaikan dan modifan beberapa hal:
-- Key untuk gerakan paddle;
-- Update untuk feature PaddleSpeedUp;
-- Apabila bola belum menyentuh paddle pada round, maka power up ExtendPaddle dan PaddleSpeedUp tidak akan memberikan efek apa-apa;
-- dll.

Dan maaf sebelumnya, sepertinya UserSettings nya ga sengaja ke-upload pas yang lalu, dikarenakan pas awal-awal kan belum aku masukin ke gitignore, trus sepertinya keburu sudah ada update, baru aku masukan ke gitignore. Akhirnya muncul terus di Github Desktop nya kalau belum dicommit.
Saya tidak berani Discard Changes juga karena takut ada perubahan yang penting tapi saya lupa.
Walaupun setelah saya lihat-lihat lagi, gitignore saya kemarin masih mengikuti yang branch Main, belum terupdate seperti yang branch Power-Up.

+++ Assets/Scripts/PUPaddleSpeedUp.cs
@@ -9,7 +9,7 @@ public class PUPaddleSpeedUp : MonoBehaviour
     public Collider2D paddleKanan;
     public Collider2D paddleKiri;
 
-    public float magnitude;
+    public int magnitude;
     public float duration;
 
     private void OnTriggerEnter2D(Collider2D collision)
